renamed proj4go->proj; moving around lots of other files and names
