+++ Proj.go
@@ -1,11 +1,13 @@
-package api
+package proj
 
 import (
 	"fmt"
 
-	"github.com/go-spatial/proj4go/core"
-	_ "github.com/go-spatial/proj4go/operations" // need to pull in the operations table
-	"github.com/go-spatial/proj4go/support"
+	"github.com/go-spatial/proj/core"
+	"github.com/go-spatial/proj/support"
+
+	// need to pull in the operations table entries
+	_ "github.com/go-spatial/proj/operations"
 )
 
 // Projector is used to perform a conversion from a lon/lat (degree) system
@@ -18,8 +20,8 @@ type Projector struct {
 	converter  core.IConvertLPToXY
 }
 
-// NewProjector creates a Projector object for the given soure and destination systems.
-func NewProjector(sourceEPSG, destEPSG int) (*Projector, error) {
+// New creates a Projector object for the given soure and destination systems.
+func New(sourceEPSG, destEPSG int) (*Projector, error) {
 	pair := pair{sourceEPSG, destEPSG}
 	str, ok := knownProjections[pair]
 	if !ok {
@@ -81,9 +83,23 @@ func (proj *Projector) Project(input []float64) ([]float64, error) {
 
 // Project performs the above functions in one step
 func Project(source int, dest int, input []float64) ([]float64, error) {
-	proj, err := NewProjector(source, dest)
+	proj, err := New(source, dest)
 	if err != nil {
 		return nil, nil
 	}
 	return proj.Project(input)
+}
+
+//---------------------------------------------------------------------
+
+type pair struct {
+	source, dest int
+}
+
+type pairTable map[pair]string
+
+var knownProjections = pairTable{
+	pair{4326, 3395}: "+proj=merc +lon_0=0 +k=1 +x_0=0 +y_0=0 +datum=WGS84", // TODO: support +units=m +no_defs
+	pair{4326, 3857}: "+proj=merc +a=6378137 +b=6378137 +lat_ts=0.0 +lon_0=0.0 +x_0=0.0 +y_0=0 +k=1.0 +units=m +nadgrids=@null +wktext +no_defs",
+	// TODO: add (3857,3395)?
 }
