+++ core/ConvertLPToXY.go
@@ -3,8 +3,8 @@ package core
 import (
 	"math"
 
-	"github.com/go-spatial/proj4go/merror"
-	"github.com/go-spatial/proj4go/support"
+	"github.com/go-spatial/proj/merror"
+	"github.com/go-spatial/proj/support"
 )
 
 // IConvertLPToXY is for 2D LP->XY conversions
