+++ core/System_test.go
@@ -3,8 +3,8 @@ package core_test
 import (
 	"testing"
 
-	"github.com/go-spatial/proj4go/core"
-	"github.com/go-spatial/proj4go/support"
+	"github.com/go-spatial/proj/core"
+	"github.com/go-spatial/proj/support"
 	"github.com/stretchr/testify/assert"
 )
