+++ merror/merror_test.go
@@ -3,7 +3,7 @@ package merror_test
 import (
 	"testing"
 
-	"github.com/go-spatial/proj4go/merror"
+	"github.com/go-spatial/proj/merror"
 	"github.com/stretchr/testify/assert"
 )
