+++ operations/august.go
@@ -3,7 +3,7 @@ package operations
 import (
 	"math"
 
-	"github.com/go-spatial/proj4go/core"
+	"github.com/go-spatial/proj/core"
 )
 
 func init() {
