+++ support/ProjectionsTable_test.go
@@ -3,7 +3,7 @@ package support_test
 import (
 	"testing"
 
-	"github.com/go-spatial/proj4go/support"
+	"github.com/go-spatial/proj/support"
 	"github.com/stretchr/testify/assert"
 )
