+++ tests/data/ellipsoid.gie
